handle huddles
The API has some strange behavior when dealing with huddles... if you let
slack change your status to "In a huddle" it does not actually change
your status (or at least not the status_text from the API), rather it
changes the huddle_state from "default_unset" to "in_a_huddle". If you
try to change your status via the API while in a huddle, it will change
the status_text, but NOT the huddle_state. If you change the status
a second time while in a huddle, it will wipe out the huddle state! So
despite being in a huddle, the API will return "default_unset" for the
huddle_state. This behavior is totally different if you uncheck
`Set my status to “:headphones: In a huddle…”` in the Audio & video
settings of the slack client, and I have not thoroughly tested that case.
But this smells like a bug in the Slack API...